+++ bluepysnap/nodes.py
@@ -128,6 +128,15 @@ class NodePopulation:
         )
 
     @property
+    def population_config(self):
+        """Access the configuration for the population.
+
+        This configuration is extended with 'components' part of the config and the filepath
+        to the h5 file containing the population.
+        """
+        return self._circuit.get_node_population_config(self.name)
+
+    @property
     def property_names(self):
         """Set of available node properties.
 
@@ -571,10 +580,10 @@ class NodePopulation:
     @cached_property
     def h5_filepath(self):
         """Get the H5 nodes file associated with population."""
-        for node_conf in self._circuit.config["networks"]["nodes"]:
-            if self.name in node_conf["populations"]:
-                return node_conf["nodes_file"]
-        raise BluepySnapError(f"h5_filepath not found for population '{self.name}'")
+        try:
+            return self.population_config["nodes_file"]
+        except KeyError as e:
+            raise BluepySnapError(f"h5_filepath not found for population '{self.name}'") from e
 
 
 class Nodes(