+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+Version v1.0.1
+--------------
+
+New Features
+~~~~~~~~~~~~
+- Access the population configs for node/edge populations via population_config property
+
 Version v1.0.0
 --------------